Rxify TestStore
